# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A Web3 backend for Sukuk (Islamic bonds) on Base Testnet, providing APIs for dApps and frontends while processing blockchain events from an indexer.",
        "title": "Sukuk POC Backend API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.sukuk-poc.com/support",
            "email": "support@sukuk-poc.com"
        },
        "license": {
            "name": "MIT",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/admin/companies": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create a new partner company (Admin only)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Create a new company",
                "parameters": [
                    {
                        "description": "Company data",
                        "name": "company",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.CreateCompanyRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Company created successfully",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request data",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin/companies/{id}": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update existing company information (admin only)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Company Management"
                ],
                "summary": "Update company",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Company ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.UpdateCompanyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Company updated",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Company not found",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin/companies/{id}/upload-logo": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Upload logo image file for a company (admin only)",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Company Management"
                ],
                "summary": "Upload company logo",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Company ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Logo image file (jpg, png, gif)",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Logo uploaded",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.FileUploadResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Company not found",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/analytics/vault/{seriesId}": {
            "get": {
                "description": "Get IDRX vault balance and utilization metrics for a specific Sukuk series",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Analytics"
                ],
                "summary": "Get vault balance",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Sukuk Series ID",
                        "name": "seriesId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Vault balance data",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.VaultBalanceResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid series ID",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Sukuk series not found",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_handlers.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string",
                    "example": "Validation failed"
                },
                "message": {
                    "type": "string",
                    "example": "Operation completed successfully"
                }
            }
        },
...
        "internal_handlers.CreateCompanyRequest": {
            "type": "object",
            "required": [
                "email",
                "name",
                "wallet_address"
            ],
            "properties": {
                "description": {
                    "type": "string",
                    "example": "Leading Indonesian sukuk issuer"
                },
                "email": {
                    "type": "string",
                    "example": "contact@sukukindonesia.com"
                },
                "industry": {
                    "type": "string",
                    "example": "Financial Services"
                },
                "name": {
                    "type": "string",
                    "example": "PT Sukuk Indonesia"
                },
                "wallet_address": {
                    "type": "string",
                    "example": "0x1234567890123456789012345678901234567890"
                },
                "website": {
                    "type": "string",
                    "example": "https://sukukindonesia.com"
                }
            }
        },
...
        "internal_handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Invalid request data"
                }
            }
        },
...
        "internal_handlers.FileUploadResponse": {
            "type": "object",
            "properties": {
                "filename": {
                    "type": "string",
                    "example": "company_logo_1.png"
                },
                "message": {
                    "type": "string",
                    "example": "File uploaded successfully"
                },
                "url": {
                    "type": "string",
                    "example": "/uploads/logos/company_logo_1.png"
                }
            }
        },
...
        "internal_handlers.UpdateCompanyRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "example": "Updated description"
                },
                "email": {
                    "type": "string",
                    "example": "newcontact@sukukindonesia.com"
                },
                "industry": {
                    "type": "string",
                    "example": "Financial Technology"
                },
                "is_active": {
                    "type": "boolean",
                    "example": true
                },
                "name": {
                    "type": "string",
                    "example": "PT Sukuk Indonesia Updated"
                },
                "website": {
                    "type": "string",
                    "example": "https://newsukukindonesia.com"
                }
            }
        },
...
        "internal_handlers.VaultBalance": {
            "type": "object",
            "properties": {
                "series_id": {
                    "type": "integer",
                    "example": 1
                },
                "series_name": {
                    "type": "string",
                    "example": "Green Sukuk Series A"
                },
                "total_supply": {
                    "type": "string",
                    "example": "1000000000000000000000000"
                },
                "utilization_rate": {
                    "type": "string",
                    "example": "0.5"
                },
                "vault_balance": {
                    "type": "string",
                    "example": "500000000000000000000000"
                }
            }
        },
        "internal_handlers.VaultBalanceResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/internal_handlers.VaultBalance"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "API key for accessing protected admin endpoints",
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        }
    }
}